Move ReferenceData to new module
